refactor: split player.ts into modular structure
